pyproject.toml
[tool.poetry]
name = "djangomongodb"
version = "0.1.0"
description = ""
authors = ["WiderAI <mukhsinmukhtorov02@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.15.1"
djongo = "1.3.3"
django = "3.0.5"
certifi = "^2024.2.2"
idna = "^3.6"
sqlparse = "0.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
